Begin new frontend design of application - add font awesome library, add new settings page and home page, add navbar and sidebar components, begin using scss variables, fix deprecation warning for scss compiler
