+++ frontend/src/assets/styles/common-variables.scss
@@ -1,0 +1,5 @@
+$primary-color: #d3d3d3;
+$secondary-color: #efefef;
+
+$nav-bar-width: 80px;
+$side-bar-width: 300px;
