+++ frontend/src/components/nav-bar/index.scss
@@ -1,0 +1,31 @@
+@use "@/assets/styles/common-variables.scss" as vars;
+
+.nav-bar {
+  position: fixed;
+  top: 0;
+  left: 0;
+
+  width: vars.$nav-bar-width;
+  height: 100vh;
+
+  padding: 30px 0 0 0;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  background-color: vars.$primary-color;
+
+  .menu-item {
+    margin: 0 0 25px 0;
+
+    cursor: pointer;
+
+    .icon {
+      text-decoration: none;
+      color: black;
+
+      font-size: 25px;
+    }
+  }
+}
