+++ frontend/src/components/side-bar/index.scss
@@ -1,0 +1,41 @@
+@use "@/assets/styles/common-variables.scss" as vars;
+
+.side-bar {
+  position: fixed;
+  top: 0;
+  left: vars.$nav-bar-width;
+
+  width: vars.$side-bar-width;
+  height: 100vh;
+
+  padding: 30px 50px 0 50px;
+
+  display: flex;
+  flex-direction: column;
+  // align-items: center;
+
+  background-color: vars.$secondary-color;
+
+  font-family: Raleway;
+  font-weight: bold;
+  text-transform: uppercase;
+
+  .title {
+    margin: 0 0 40px 0;
+    font-size: 22px;
+    font-weight: bolder;
+  }
+
+  .header {
+    margin: 0 0 25px 0;
+    font-size: 18px;
+  }
+
+  .sidebar-item {
+    margin: 0 0 25px 0;
+
+    font-size: 14px;
+
+    cursor: pointer;
+  }
+}
