+++ frontend/src/fontAwesome.ts
@@ -1,0 +1,5 @@
+import { library } from "@fortawesome/fontawesome-svg-core";
+import { faListCheck, faGear } from "@fortawesome/free-solid-svg-icons"; // Solid icons
+
+library.add(faListCheck);
+library.add(faGear);
