+++ frontend/src/main.tsx
@@ -1,6 +1,9 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
+
 import "./index.scss";
+import "./fontAwesome.ts";
+
 import App from "./App";
 
 createRoot(document.getElementById("root")!).render(
