+++ frontend/src/pages/home/dep_index.scss
